homeassistant/componen: Add daikin p1/p2 module
A new component for Daikin heat-pumps. Inspired by [1].

Interfaces with a homebrew microcontroller that communicates
over the serial. The device is supposed to relay P1/P2 received packets
over the serial as ASCII hex data. Each line represents a single packet.

The P1/P2 bus is a two wire powerline bus, derived from the Japanese
Home Bus System (published 1986 as the ET-2101). It runs at 9600baud and
uses amplitude modulation on the DC signal that poweres the bus.
It is a half-duplex and a polled bus. The heat-pump always initiates the
communication and probes for additional nodes on the bus.

This component decodes the hex data and exposes it as sensors.

1: https://github.com/Arnold-n/P1P2MQTT

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>